[02旅游景点智能推荐平台]  feat: 更新景点相关DTO和VO，优化数据结构，增强功能支持

+++ projects/02travel_recommend_system/travel-web/src/main/java/com/travel/advisor/dto/scenic/ScenicCreateDTO.java
@@ -7,6 +7,9 @@ import lombok.Data;
 import java.math.BigDecimal;
 import java.util.List;
 
+/**
+ * 景点创建 DTO
+ */
 @Data
 public class ScenicCreateDTO {
 
@@ -39,7 +42,7 @@ public class ScenicCreateDTO {
 
     private String category;
 
-    private BigDecimal ratingScore;
+    private Double ratingScore;
 
     private String bestSeason;
